fix: UserStatus 관련 정규화 코드 정리
