+++ apps/mac/src/renderer/src/entities/competition/index.ts
@@ -34,6 +34,8 @@ export type {
   PostBattleRequest,
 } from "./model/rival-competition/battle.types";
 export type { MyRivalsRequest, MyRivalsResponse } from "./model/rival-competition/myRivals.types";
+export { USER_STATUS_LABELS } from "./model/rival-competition/userStatus";
+export type { UserStatus } from "./model/rival-competition/userStatus";
 export type {
   CategoryType,
   PeriodType,
