+++ apps/mac/src/renderer/src/shared/lib/index.ts
@@ -14,7 +14,6 @@ export { useRival } from "./useRival";
 export { buildPaddedStreak } from "./buildPaddedStreaks";
 export { getGrowthInfo } from "./getGrowthInfo";
 export { getCountLabel } from "./getCountLabel";
-export type { MyRivalItem } from "./useRival";
 export { resolveUsingApp } from "./resolveUsingApp";
 export { useRealtimeRivalActiveTime } from "./useRealtimeRivalActiveTime";
 export { resolveProfileDecorations } from "./profileDecorations";
