+++ apps/mac/src/renderer/src/shared/ui/index.ts
@@ -17,7 +17,6 @@ export { RivalsManagementDialog, DeleteRivalsConfirmDialog } from "./rival-manag
 export { SearchInput } from "./search-input";
 export { SlideSelector } from "./slide-selector";
 export type { SlideSelectorProps } from "./slide-selector";
-export { MyRivalUsers } from "./myrival-users/MyRivalUsers";
 export { UserRanking } from "./home-ranking-user/UserRanking";
 export { GrowthRate } from "./growth-rate/GrowthRate";
 export { RivalCompetitionLineChart } from "./rival-compete-chart/RivalCompeteLineChart";
